+++ include/KeyFrame.h
@@ -58,7 +58,7 @@ class KeyFrame
     {
         ar & mnId;
         ar & const_cast<long unsigned int&>(mnFrameId);
-        ar & const_cast<double&>(mTimeStamp);
+        ar & const_cast<long double&>(mTimeStamp);
         // Grid
         ar & const_cast<int&>(mnGridCols);
         ar & const_cast<int&>(mnGridRows);
@@ -312,7 +312,7 @@ public:
     long unsigned int mnId;
     const long unsigned int mnFrameId;
 
-    const double mTimeStamp;
+    const long double mTimeStamp;
 
     // Grid (to speed up feature matching)
     const int mnGridCols;
